// pdf-editor/src/code/contractManagement/builder/types.ts
export interface PageDimension {
  pageWidth: number;
  pageHeight: number;
}